Refactor: Move quiz-related files to dedicated quiz package structure
- Created new quiz package with controller, dto, model, repository, and service folders
- Moved QuizController to quiz.controller package
- Moved QuizService to quiz.service package
- Moved quiz DTOs (QuestionResponseDto, QuizConfigurationDto, QuizResponseDto) to quiz.dto package
- Moved quiz models (Question, Quiz) to quiz.model package
- Updated import statements in GeminiService to reference new quiz.dto.QuizConfigurationDto
- Deleted old quiz files from main package structure
- Maintained separation between quiz-specific and shared functionality (GeminiService, GeminiResponseDto remain in main package)
- All files compile successfully after refactoring
