+++ backend/src/main/java/com/bestgroup/HomeEntertAInment/quiz/dto/QuizConfigurationDto.java
@@ -1,4 +1,4 @@
-package com.bestgroup.HomeEntertAInment.dto;
+package com.bestgroup.HomeEntertAInment.quiz.dto;
 
 import lombok.Data;
 import lombok.NoArgsConstructor;
